Add more bins to miles downstream, add keyboard events for unit search, minor fixes

+++ analysis/prep/barriers/prep_dams.py
@@ -351,7 +351,7 @@ df["HeightClass"] = np.asarray(
 df["PassageFacilityClass"] = np.uint8(1)
 df.loc[(df.PassageFacility > 0) & (df.PassageFacility != 9), "PassageFacilityClass"] = 2
 
-df["FeasibilityClass"] = df.Feasibility.map(FEASIBILITY_TO_DOMAIN)
+df["FeasibilityClass"] = df.Feasibility.map(FEASIBILITY_TO_DOMAIN).astype("uint8")
 
 
 # Convert BarrierSeverity to a domain
